requirements.txt
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.6.2
Werkzeug==3.1.4
